src/inference/requirements.txt
torchdata
hydra-core
omegaconf
tqdm
accelerate
wandb
einops
diffusers==0.36.0
huggingface_hub==0.36.0
transformers==4.57.6
alias_free_torch
h5py
torchaudio
soundfile
tensorboard
swanlab
pydantic==2.11.7   
#pydantic==2.12.2